Use test naming convention for test files
