+++ src/core/kilocode/agent-manager/__tests__/AgentManagerProvider.test.ts
@@ -1,11 +1,12 @@
 import { describe, it, expect, vi, beforeEach, afterEach, type Mock } from "vitest"
 import { EventEmitter } from "node:events"
 
+const MOCK_CLI_PATH = "/mock/path/to/kilocode"
+
 let AgentManagerProvider: typeof import("../AgentManagerProvider").AgentManagerProvider
 
-describe("AgentManagerProvider - CLI backend", () => {
+describe("AgentManagerProvider CLI spawning", () => {
 	let provider: InstanceType<typeof AgentManagerProvider>
-
 	const mockContext = { extensionUri: {}, extensionPath: "" } as any
 	const mockOutputChannel = { appendLine: vi.fn() } as any
 
@@ -18,8 +19,13 @@ describe("AgentManagerProvider - CLI backend", () => {
 		vi.doMock("vscode", () => ({
 			workspace: { workspaceFolders: [mockWorkspaceFolder] },
 			window: mockWindow,
-			Uri: { joinPath: () => undefined },
+			env: { openExternal: vi.fn() },
+			Uri: { parse: vi.fn(), joinPath: vi.fn() },
 			ViewColumn: { One: 1 },
+		}))
+
+		vi.doMock("../../../../utils/fs", () => ({
+			fileExistsAtPath: vi.fn().mockResolvedValue(false),
 		}))
 
 		class TestProc extends EventEmitter {
@@ -30,9 +36,12 @@ describe("AgentManagerProvider - CLI backend", () => {
 		}
 
 		const spawnMock = vi.fn(() => new TestProc())
-		const execSyncMock = vi.fn(() => "kilocode")
+		const execSyncMock = vi.fn(() => MOCK_CLI_PATH)
 
-		vi.doMock("node:child_process", () => ({ spawn: spawnMock, execSync: execSyncMock }))
+		vi.doMock("node:child_process", () => ({
+			spawn: spawnMock,
+			execSync: execSyncMock,
+		}))
 
 		const module = await import("../AgentManagerProvider")
 		AgentManagerProvider = module.AgentManagerProvider
@@ -44,17 +53,14 @@ describe("AgentManagerProvider - CLI backend", () => {
 	})
 
 	it("spawns kilocode without shell interpolation for prompt arguments", async () => {
-		const childProcess = await (provider as any).startAgentSession('echo "$(whoami)"')
+		await (provider as any).startAgentSession('echo "$(whoami)"')
 
 		const spawnMock = (await import("node:child_process")).spawn as unknown as Mock
 		expect(spawnMock).toHaveBeenCalledTimes(1)
 		const [cmd, args, options] = spawnMock.mock.calls[0] as unknown as [string, string[], Record<string, unknown>]
-		expect(cmd).toBe("kilocode")
+		expect(cmd).toBe(MOCK_CLI_PATH)
 		expect(args[args.length - 1]).toBe('echo "$(whoami)"')
 		expect(options?.shell).not.toBe(true)
-
-		// Ensure the child process is cleaned up
-		;(childProcess as any)?.kill?.("SIGTERM")
 	})
 
 	it("flushes buffered CLI output on process exit", async () => {
@@ -246,5 +252,51 @@ describe("AgentManagerProvider - CLI backend", () => {
 		expect(messages).toHaveLength(1)
 		expect(messages?.[0].text).toBe("final")
 		expect(messages?.[0].partial).toBe(false)
+	})
+
+	describe("dispose behavior", () => {
+		it("kills all running processes on dispose", async () => {
+			await (provider as any).startAgentSession("session 1")
+			await (provider as any).startAgentSession("session 2")
+
+			const spawnMock = (await import("node:child_process")).spawn as unknown as Mock
+			const proc1 = spawnMock.mock.results[0].value
+			const proc2 = spawnMock.mock.results[1].value
+
+			expect((provider as any).processes.size).toBe(2)
+
+			provider.dispose()
+
+			expect(proc1.kill).toHaveBeenCalledWith("SIGTERM")
+			expect(proc2.kill).toHaveBeenCalledWith("SIGTERM")
+			expect((provider as any).processes.size).toBe(0)
+		})
+
+		it("clears all timeouts on dispose", async () => {
+			await (provider as any).startAgentSession("session with timeout")
+
+			expect((provider as any).timeouts.size).toBe(1)
+
+			provider.dispose()
+
+			expect((provider as any).timeouts.size).toBe(0)
+		})
+	})
+
+	describe("hasRunningSessions", () => {
+		it("returns false when no sessions exist", () => {
+			expect((provider as any).hasRunningSessions()).toBe(false)
+		})
+
+		it("returns true when a session is running", async () => {
+			await (provider as any).startAgentSession("running")
+			expect((provider as any).hasRunningSessions()).toBe(true)
+		})
+
+		it("returns count of running sessions", async () => {
+			await (provider as any).startAgentSession("running 1")
+			await (provider as any).startAgentSession("running 2")
+			expect((provider as any).getRunningSessionCount()).toBe(2)
+		})
 	})
 })
