+++ src/core/kilocode/agent-manager/__tests__/AgentRegistry.test.ts
