+++ src/core/kilocode/agent-manager/__tests__/CliPathResolver.test.ts
